Increase the accuracy of position fields from Float to Double
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@20420 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: c985fa93d635a7494df364ef501f6cc1e2f8439a

+++ server/wrapper/ivef/de.uniol.inf.is.odysseus.wrapper.ivef/src/de/uniol/inf/is/odysseus/wrapper/ivef/parser/IVEF_0_1_5_Parser.java
@@ -264,11 +264,11 @@ public class IVEF_0_1_5_Parser extends DefaultHandler {
             String value = atts.getValue(i);
 
             if (key == "Lat") {
-            	Float val = Float.parseFloat(value);
+            	Double val = Double.parseDouble(value);
                 obj.setLat(val);
             }
             else if (key == "Long") {
-            	Float val = Float.parseFloat(value);
+            	Double val = Double.parseDouble(value);
                 obj.setLong(val);
             }
         }
